requirements.txt
asgiref==3.8.1
autopep8==2.3.1
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.4.0
pycodestyle==2.12.1
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
